Remove defunt styles and move NewButton into Timeline
