+++ src/components/OrderedList/index.js
@@ -3,6 +3,5 @@ export { default as Item } from './Item';
 export { default as Row } from './Row';
 export { default as Group } from './Group';
 export { default as ExpandableItem } from './ExpandableItem';
-export { default as NewButton } from './NewButton';
 export { default as Handle } from './Handle';
 export { default as DeleteButton } from './DeleteButton';
