+++ src/components/UnorderedList/index.js
@@ -3,6 +3,5 @@ export { default } from './FieldArrayAdapter';
 export { default as UnorderedList } from './UnorderedList';
 export { default as Item } from '../OrderedList/Item';
 export { default as ExpandableItem } from '../OrderedList/ExpandableItem';
-export { default as NewButton } from '../OrderedList/NewButton';
 export { default as Handle } from '../OrderedList/Handle';
 export { default as DeleteButton } from '../OrderedList/DeleteButton';
