+++ src/styles/components/_list.scss
@@ -5,6 +5,5 @@
 @import './list/list';
 @import './list/list-controls';
 @import './list/list-sort-control';
-@import './list/new-button';
 @import './list/list-row';
 @import './list/list-group';
