+++ src/styles/components/_timeline.scss
@@ -1,4 +1,3 @@
 @import 'timeline/timeline';
-@import 'timeline/add-new';
-@import 'timeline/insert-stage';
+@import 'timeline/new-button';
 @import 'timeline/stage';
