+++ src/styles/components/stage-editor/section/_content-items.scss
@@ -2,8 +2,10 @@
   &__controls {
     margin-top: unit(2);
 
-    .button:first-child {
-      margin-left: 0;
+    .button {
+      &:first-child {
+        margin-left: 0;
+      }
     }
   }
