--- src/styles/components/timeline/_add-new.scss
@@ -1,49 +1,0 @@
-$component-name: timeline-add-new;
-
-.#{$component-name} {
-  padding: unit(5) 0;
-  text-align: center;
-
-  &__button {
-    cursor: pointer;
-    background-color: transparent;
-    border-radius: 5px;
-    transform: translateZ(0) scale(.75);
-    transition: background-color var(--animation-easing) var(--animation-duration-standard),
-    transform var(--animation-easing) var(--animation-duration-standard);
-    padding: 15px 60px;
-    border: 0;
-
-    &-label {
-      transition: opacity var(--animation-easing) var(--animation-duration-standard);
-      opacity: 0;
-      color: var(--text);
-    }
-
-    &-icon {
-      transform: translateX(6px);
-      width: 80px !important; // sass-lint:disable-line no-important
-      height: 80px !important; // sass-lint:disable-line no-important
-    }
-
-    &:hover {
-      background-color: var(--background);
-      transform: translateZ(0) scale(1);
-
-      .#{$component-name}__button-label {
-        opacity: 1;
-      }
-    }
-
-    &:focus {
-      outline: none;
-    }
-  }
-
-  &:last-child {
-    &::after {
-      height: 50%;
-    }
-  }
-
-}
