--- src/styles/components/timeline/_insert-stage.scss
@@ -1,4 +1,0 @@
-@import './insert-stage/insert-stage';
-@import './insert-stage/filter';
-@import './insert-stage/option-grid';
-@import './insert-stage/guidance';
