+++ src/styles/components/timeline/_new-button.scss
