--- src/styles/components/timeline/insert-stage/_filter.scss
@@ -1,45 +1,0 @@
-$component-name: timeline-insert-stage-filter;
-
-.#{$component-name} {
-  display: flex;
-  padding: unit(2) unit(12);
-  background-color: var(--color-navy-taupe);
-  color: var(--text-light);
-
-  input {
-    color: var(--text);
-  }
-
-  &__option {
-    margin-right: unit(2);
-    width: 300px;
-
-    .form-field-container {
-      margin-bottom: 0;
-
-      .form-field {
-        margin-bottom: 0;
-      }
-    }
-
-    h4 {
-      font-family: var(--body-font-family);
-      margin-bottom: unit(1);
-      font-size: .8rem;
-    }
-  }
-
-  &__option--last {
-    flex: 1 0 auto;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-  }
-
-  &__reset {
-    text-decoration: underline;
-    font-family: var(--body-font-family);
-    margin-bottom: unit(1);
-    cursor: pointer;
-  }
-}
