--- src/styles/components/timeline/insert-stage/_option-grid.scss
@@ -1,45 +1,0 @@
-$stage-columns: 3;
-$stage-gap: unit(1);
-$stage-width: calc(#{100% / $stage-columns} - #{($stage-columns - 1) * $stage-gap / $stage-columns});
-$component-name: timeline-insert-stage-option-grid;
-
-.#{$component-name} {
-  display: grid;
-  padding: unit(2);
-  grid-template-columns: repeat($stage-columns, $stage-width);
-  grid-gap: $stage-gap;
-  grid-auto-rows: minmax(100px, auto);
-
-  &__option {
-    cursor: pointer;
-    background: var(--background-light);
-    padding: unit(2);
-
-    &:hover {
-      .#{$component-name}__preview {
-        border-color: var(--architect-intent);
-      }
-    }
-
-    h3 {
-      height: 2.2em;
-      margin-bottom: unit(1);
-    }
-  }
-
-  &__screen,
-  &__description {
-    width: 100%;
-  }
-
-  &__description {
-    height: 100%;
-  }
-
-  &__preview {
-    border: 2px solid transparent;
-    overflow: hidden;
-    transition: border-color var(--animation-duration-standard) var(--animation-easing);
-    position: relative;
-  }
-}
